test/vmtests/requirements.txt
docker == 7.1.0
libvirt-python == 10.9.0
paramiko == 3.5.0
pytest == 8.3.3
PyYAML == 6.0.2
types-docker == 7.1.0.20240827
